neurok-tts/requirements.txt
matplotlib==3.10.1
torch==2.6.0
numpy==2.1.3
PyYAML==6.0.2
pydantic-settings==2.8.1
pydantic==2.10.6
pika==1.3.2
scipy==1.15.2
nltk==3.9.1
phonemizer==3.3.0
librosa==0.11.0
soundfile==0.13.1
tqdm==4.67.1
estnltk==1.7.4
tensorflow==2.19.0
torchaudio==2.6.0
torchvision==0.21.0
python-dotenv==1.0.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
tf-keras==2.19.0
nvidia-cudnn-cu11==9.3.0.75